abstracted person viewed to eventUsageLogic
